* Moved code used to spawn new processes into rUtilAnts/ForeignProcess. * Code clean-up. !!! Not finished yet. * Prepared SQL code to be executed in a separate process if needed. !!! Not tested yet. * Added :GemHomeDir and :RubyGemsLibDir options to SlaveProduct/Redmine (regression updated). * WEACE environment files are not overwritten if already present (they could be adapted by the user). * Adapted SlaveProduct/Redmine's logProduct method to the new SQL framework.
git-svn-id: https://weacemethod.svn.sourceforge.net/svnroot/weacemethod@178 5f6fdd18-d50c-4247-9f77-f36982294e7a

+++ WEACEToolkit/lib/WEACEToolkit/Install/WEACE_InstallCommon.rb
@@ -52,14 +52,18 @@ module WEACEInstall
       if (lRUALibDir == nil)
         logBug 'Unable to find rUtilAnts library directory among Rubygems or $LOAD_PATH.'
       else
-        File.open(@WEACEEnvFile, 'w') do |oFile|
-          oFile << "\# This file is meant to be required by every script that needs to setup WEACE environment before loading and using WEACE Toolkit libraries.
+        if (File.exists?(@WEACEEnvFile))
+          logWarn "Environment file #{@WEACEEnvFile} already exists. Will not overwrite it."
+        else
+          File.open(@WEACEEnvFile, 'w') do |oFile|
+            oFile << "\# This file is meant to be required by every script that needs to setup WEACE environment before loading and using WEACE Toolkit libraries.
 
 \# Path containing rUtilAnts library (can be path to RubyGems)
 $LOAD_PATH << '#{File.expand_path(lRUALibDir)}'
 \# Path containing WEACE Toolkit library
 $LOAD_PATH << '#{File.expand_path(File.dirname(@WEACELibDir))}'
 "
+          end
         end
       end
     end
